requirements.txt
asgiref==3.7.2
boto3==1.34.58
botocore==1.34.58
dj-database-url==2.1.0
Django==5.0.3
django-ckeditor==6.7.1
django-environ==0.11.2
django-heroku==0.3.1
django-js-asset==2.2.0
django-multiupload==0.6.1
django-storages==1.14.2
gunicorn==21.2.0
jmespath==1.0.1
packaging==23.2
python-dateutil==2.9.0.post0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.0.7
whitenoise==6.6.0
